pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "appaveli-codemind"
version = "0.1.0"
description = "AI-powered code refactoring, security scanning, and generation"
readme = "README.md"
authors = [
    { name = "AppaveliTech Solutions", email = "support@appaveli.com" }
]
requires-python = ">=3.8"
license = "MIT"  # ✅ SPDX format instead of [text = "MIT"]
keywords = ["ai", "code-generation", "security", "refactoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Security"
]
dependencies = [
    "openai>=1.12.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0",
    "isort>=5.12.0"
]

[project.scripts]
appaveli-codemind = "appaveli_codemind.main:main"

[tool.setuptools]
packages = ["appaveli_codemind"]  # ✅ This avoids the "multiple top-level packages" error